+++ src/main/java/com/irontechspace/dynamicdq/executor/export/ExportExcelService.java
@@ -9,9 +9,7 @@ import com.irontechspace.dynamicdq.configurator.query.model.QueryConfig;
 import com.irontechspace.dynamicdq.configurator.query.model.QueryField;
 import com.irontechspace.dynamicdq.executor.file.FileService;
 import lombok.extern.log4j.Log4j2;
-import org.apache.poi.hssf.util.HSSFColor;
 import org.apache.poi.ss.usermodel.*;
-import org.apache.poi.ss.util.CellRangeAddress;
 import org.apache.poi.xssf.usermodel.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mock.web.MockMultipartFile;
@@ -20,6 +18,8 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
@@ -60,43 +60,54 @@ public class ExportExcelService {
     }
 
     public JsonNode addTableExcel(String configName, UUID userId, List<String> userRoles, JsonNode body) {
-        QueryConfig config = queryConfigService.getByName(configName, userId, userRoles);
+
+        List<ExcelCol> fields = new ArrayList<>();
+        if(configName != null) {
+            QueryConfig config = queryConfigService.getByName(configName, userId, userRoles);
+            fields = config.getFields().stream().filter(QueryField::getVisible).map(ExcelCol::new).collect(Collectors.toList());
+        } else if (body.get("fields") != null && body.get("fields").isArray()) {
+            ArrayNode fieldsArray = (ArrayNode) body.get("fields");
+            for(JsonNode jsonField : fieldsArray)
+                fields.add(new ExcelCol(jsonField));
+        }
+
         try {
             XSSFWorkbook workbook = OBJECT_MAPPER.treeToValue(body.get("file"), XSSFWorkbook.class);
             Sheet sheet = workbook.getSheet(body.get("sheet").asText());
             ArrayNode data = (ArrayNode) body.get("data");
+            boolean hiddenHeader = !body.path("hiddenHeader").isMissingNode() && body.get("hiddenHeader").asBoolean();
+            float headerHeight = body.path("headerHeight").isMissingNode() ? -1 : body.get("headerHeight").floatValue();
+            float rowHeight = body.path("rowHeight").isMissingNode() ? -1 : body.get("rowHeight").floatValue();
             int startRowIndex = body.get("startCell").get("row").asInt();
             int startColIndex = body.get("startCell").get("col").asInt();
-            List<QueryField> fields = config.getFields().stream().filter(QueryField::getVisible).collect(Collectors.toList());
 
-            Row headers = sheet.createRow(startRowIndex++);
-            for (int colIndex = startColIndex; colIndex < fields.size(); colIndex++) {
-                QueryField field = fields.get(colIndex);
-                Cell cell = headers.createCell(colIndex);
-//                XSSFFont font = workbook.createFont();
-//                font.setBold(true);
-                CellStyle style = workbook.createCellStyle();
-//                style.setFont(font);
-                style.setAlignment(HorizontalAlignment.valueOf(field.getAlign().toUpperCase()));
-                style.setVerticalAlignment(VerticalAlignment.CENTER);
-                sheet.setColumnWidth(colIndex, pixel2WidthUnits(field.getWidth().intValue()));
-                cell.setCellStyle(style);
-                cell.setCellValue(field.getHeader());
+            if(hiddenHeader) {
+                for (int colIndex = startColIndex; colIndex < fields.size(); colIndex++) {
+                    sheet.setColumnWidth(colIndex, pixel2WidthUnits(fields.get(colIndex).getWidth().intValue()));
+                }
+            } else {
+                Row headers = Optional.ofNullable(sheet.getRow(startRowIndex)).orElse(sheet.createRow(startRowIndex));
+                startRowIndex++;
+                headers.setHeightInPoints(headerHeight);
+                for (int colIndex = startColIndex; colIndex < fields.size(); colIndex++) {
+                    ExcelCol field = fields.get(colIndex);
+                    sheet.setColumnWidth(colIndex, pixel2WidthUnits(field.getWidth().intValue()));
+
+                    Cell cell = headers.createCell(colIndex);
+                    cell.setCellStyle(getHeaderStyle(workbook, field));
+                    cell.setCellValue(field.getHeader());
+                }
             }
 
-//            int lastRowIndex = startRowIndex - 1 + data.size();
             for (int dataIndex = 0; dataIndex < data.size(); dataIndex++) {
                 JsonNode rowData = data.get(dataIndex);
-                Row row = sheet.createRow(startRowIndex + dataIndex);
+                Row row = Optional.ofNullable(sheet.getRow(startRowIndex + dataIndex)).orElse(sheet.createRow(startRowIndex + dataIndex));
+                row.setHeightInPoints(rowHeight);
                 for (int colIndex = startColIndex; colIndex < fields.size(); colIndex++) {
-                    QueryField field = fields.get(colIndex);
+                    ExcelCol field = fields.get(colIndex);
                     Cell cell = row.createCell(colIndex);
-                    CellStyle style = workbook.createCellStyle();
-                    style.setAlignment(HorizontalAlignment.valueOf(field.getAlign().toUpperCase()));
-                    style.setVerticalAlignment(VerticalAlignment.CENTER);
-//                    sheet.setColumnWidth(colIndex, field.getWidth().intValue());
-                    cell.setCellStyle(style);
-                    cell.setCellValue(rowData.get(field.getAliasOrName()).asText());
+                    cell.setCellStyle(getCellStyle(workbook, field));
+                    setCellValue(cell, field, rowData);
                 }
             }
 
@@ -107,9 +118,35 @@ public class ExportExcelService {
         }
     }
 
-//    private void addRowExcel() {
-//
-//    }
+    private XSSFFont getHeaderFont(XSSFWorkbook workbook) {
+        XSSFFont font = workbook.createFont();
+        font.setBold(true);
+        return font;
+    }
+
+    private CellStyle getHeaderStyle(XSSFWorkbook workbook, ExcelCol field) {
+        CellStyle style = getCellStyle(workbook, field);
+        style.setFont(getHeaderFont(workbook));
+        return style;
+    }
+
+    private CellStyle getCellStyle(XSSFWorkbook workbook, ExcelCol field) {
+        CellStyle style = workbook.createCellStyle();
+        if(Arrays.asList("date", "time", "timestamp").contains(field.getTypeData())) {
+            style.setDataFormat(workbook.getCreationHelper().createDataFormat().getFormat(field.getDateFormat()));
+        }
+        style.setAlignment(HorizontalAlignment.valueOf(field.getAlign().toUpperCase()));
+        style.setVerticalAlignment(VerticalAlignment.CENTER);
+        return style;
+    }
+
+    private void setCellValue(Cell cell, ExcelCol field, JsonNode rowData){
+        if(Arrays.asList("date", "time", "timestamp").contains(field.getTypeData())) {
+            cell.setCellValue(LocalDateTime.parse(rowData.get(field.getName()).asText(), DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ssXXX")));
+        } else {
+            cell.setCellValue(rowData.get(field.getName()).asText());
+        }
+    }
 
     public Boolean addRowExcel(JsonNode body) {
         try {
@@ -119,11 +156,9 @@ public class ExportExcelService {
             JsonNode data = body.get("data");
             AtomicInteger colIndex = new AtomicInteger(body.get("startCell").get("col").asInt());
             data.fields().forEachRemaining(field -> {
-                sheet.autoSizeColumn(colIndex.get());
                 addCell(row, colIndex.getAndIncrement(), field.getValue().asText());
             });
-//            if(sheet.getMergedRegion(0) != null)
-//            sheet.addMergedRegion(new CellRangeAddress(0, 0, 0, 3));
+            // sheet.addMergedRegion(new CellRangeAddress(0, 0, 0, 3));
             return true;
         } catch (IOException e) {
             e.printStackTrace();